+++ src/main/java/com/domain/usecase/cashflow/IGetCashFlowByConceptUseCase.java
@@ -3,6 +3,6 @@ package com.domain.usecase.cashflow;
 import com.domain.entities.CashFlowEntity;
 
 public interface IGetCashFlowByConceptUseCase {
-  CashFlowEntity getCashFlowByConcept(Integer conceptId);
+  CashFlowEntity getCashFlowByConcept(Long conceptId);
 
 }
