+++ src/main/java/com/domain/usecase/cashflow/IUpsertCashFlowUseCase.java
@@ -3,6 +3,6 @@ package com.domain.usecase.cashflow;
 import com.domain.entities.CashFlowEntity;
 
 public interface IUpsertCashFlowUseCase {
-  CashFlowEntity upsertCashFlow(Integer conceptId);
+  CashFlowEntity upsertCashFlow(Long conceptId);
 
 }
